deal with tcp seq loop back
